+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 2.7.0 - 2019-11-07
 
 * Added: Alias cred to the Use-VaultCredential
 * Changed: .NET Library structure and unit tests